started to make robot move during autonomous mode
Made moveRobotCurrentNew which moves the robot to a point based on a
current (x,y) coordinate to a new coordinate. Created a class to test
servos with. Started to create what the robot does based on the
autonomous questions. started testing moveRobotCurrentNew - told it to
go a foot, it went 6 inches